// packages/core/src/styleProps/types.ts
import { Color, FontFamily, TypographyType } from '../theme/types';

export type BaseStyleProps = {
  w?: string;
  maxWidth?: string;
  h?: string;
  maxHeight?: string;
  p?: number;
  px?: number;
  py?: number;
  pt?: number;
  pr?: number;
  pb?: number;
  pl?: number;
  m?: number;
  mx?: number;
  my?: number;
  mt?: number;
  mr?: number;
  mb?: number;
  ml?: number;
  color?: Color;
  bg?: Color;
  typography?: TypographyType;
  fontSize?: string;
  fontFamily?: FontFamily;
  fontWeight?: 400 | 500 | 600 | 700 | 800 | 900;
  border?: boolean;
  borderTop?: boolean;
  borderRight?: boolean;
  borderBottom?: boolean;
  borderLeft?: boolean;
  borderColor?: Color;
  borderWidth?: number;
  rounded?: boolean;
  borderRadius?: number;
  clickable?: boolean;
};

export type BaseStyleProp = keyof BaseStyleProps;